+++ test/src/e2e_vm_tests/test_programs/should_fail/type_check_analyze_errors/snapshot.toml
@@ -1,0 +1,1 @@
+
