+++ test/src/e2e_vm_tests/test_programs/should_fail/type_check_analyze_errors/src/main.sw
@@ -1,7 +1,6 @@
 script;
 
 fn main() {
-    // u8
     let _a = 0x100;
     Vec::<u8>::new().push(_a);
