--- test/src/e2e_vm_tests/test_programs/should_fail/type_check_analyze_errors/test.toml
@@ -1,2 +1,0 @@
-category = "fail"
-snapshot = true
