--- test/src/lib.rs
@@ -1,163 +1,0 @@
-pub mod e2e_vm_tests;
-mod ir_generation;
-pub mod reduced_std_libs;
-pub mod test_consistency;
-
-use anyhow::Result;
-use clap::Parser;
-use e2e_vm_tests::{FilterConfig, RunConfig};
-use forc::cli::shared::{PrintAsmCliOpt, PrintIrCliOpt};
-use forc_tracing::init_tracing_subscriber;
-use std::str::FromStr;
-use sway_core::{BuildTarget, PrintAsm, PrintIr};
-use tracing::Instrument;
-
-#[derive(Parser)]
-pub struct Cli {
-    /// Only run tests matching this regex
-    #[arg(value_name = "REGEX")]
-    pub include: Option<regex::Regex>,
-
-    /// Exclude tests matching this regex
-    #[arg(long, short, value_name = "REGEX")]
-    pub exclude: Option<regex::Regex>,
-
-    /// Skip all tests until a test matches this regex
-    #[arg(long, short, value_name = "REGEX")]
-    pub skip_until: Option<regex::Regex>,
-
-    /// Only run tests with ABI JSON output validation
-    #[arg(long, visible_alias = "abi")]
-    pub abi_only: bool,
-
-    /// Only run tests with no core dependencies
-    #[arg(long, visible_alias = "exclude_core")]
-    pub exclude_core: bool,
-
-    /// Only run tests with no std dependencies
-    #[arg(long, visible_alias = "exclude_std")]
-    pub exclude_std: bool,
-
-    /// Only run tests that deploy contracts
-    #[arg(long, visible_alias = "contract")]
-    pub contract_only: bool,
-
-    /// Only run the first test
-    #[arg(long, visible_alias = "first")]
-    pub first_only: bool,
-
-    /// Only run snapshot test
-    #[arg(long, visible_alias = "snapshot")]
-    pub snapshot_only: bool,
-
-    /// Print out warnings, errors, and output of print options
-    #[arg(long, env = "SWAY_TEST_VERBOSE")]
-    pub verbose: bool,
-
-    /// Compile sway code in release mode
-    #[arg(long)]
-    pub release: bool,
-
-    /// Intended for use in `CI` to ensure test lock files are up to date
-    #[arg(long)]
-    pub locked: bool,
-
-    /// Build target
-    #[arg(long, visible_alias = "target")]
-    pub build_target: Option<String>,
-
-    /// Disable the "new encoding" feature
-    #[arg(long)]
-    pub no_encoding_v1: bool,
-
-    /// Update all output files
-    #[arg(long)]
-    pub update_output_files: bool,
-
-    /// Print out the specified IR (separate options with comma), if the verbose option is on
-    #[arg(long, num_args(1..=18), value_parser = clap::builder::PossibleValuesParser::new(PrintIrCliOpt::cli_options()))]
-    pub print_ir: Option<Vec<String>>,
-
-    /// Print out the specified ASM (separate options with comma), if the verbose option is on
-    #[arg(long, num_args(1..=5), value_parser = clap::builder::PossibleValuesParser::new(&PrintAsmCliOpt::CLI_OPTIONS))]
-    pub print_asm: Option<Vec<String>>,
-
-    /// Print out the final bytecode, if the verbose option is on
-    #[arg(long)]
-    pub print_bytecode: bool,
-}
-
-pub fn configs_from_cli(cli: &Cli) -> (FilterConfig, RunConfig) {
-    let filter_config = FilterConfig {
-        include: cli.include.clone(),
-        exclude: cli.exclude.clone(),
-        skip_until: cli.skip_until.clone(),
-        abi_only: cli.abi_only,
-        exclude_core: cli.exclude_core,
-        exclude_std: cli.exclude_std,
-        contract_only: cli.contract_only,
-        first_only: cli.first_only,
-        snapshot_only: cli.snapshot_only,
-    };
-    let build_target = match &cli.build_target {
-        Some(target) => match BuildTarget::from_str(target.as_str()) {
-            Ok(target) => target,
-            _ => panic!("unexpected build target"),
-        },
-        None => BuildTarget::default(),
-    };
-    let run_config = RunConfig {
-        locked: cli.locked,
-        verbose: cli.verbose,
-        release: cli.release,
-        build_target,
-        experimental: sway_core::ExperimentalFlags {
-            new_encoding: !cli.no_encoding_v1,
-        },
-        update_output_files: cli.update_output_files,
-        print_ir: cli
-            .print_ir
-            .as_ref()
-            .map_or(PrintIr::default(), |opts| PrintIrCliOpt::from(opts).0),
-        print_asm: cli
-            .print_asm
-            .as_ref()
-            .map_or(PrintAsm::default(), |opts| PrintAsmCliOpt::from(opts).0),
-        print_bytecode: cli.print_bytecode,
-    };
-
-    (filter_config, run_config)
-}
-
-pub async fn run(cli: Cli) -> Result<()> {
-    init_tracing_subscriber(Default::default());
-
-    let (filter_config, run_config) = configs_from_cli(&cli);
-
-    // Check that the tests are consistent
-    test_consistency::check()?;
-
-    // Create reduced versions of the `std` library.
-    reduced_std_libs::create()?;
-
-    // Run E2E tests
-    e2e_vm_tests::run(&filter_config, &run_config)
-        .instrument(tracing::trace_span!("E2E"))
-        .await?;
-
-    // Run IR tests
-    if !filter_config.first_only && !filter_config.snapshot_only {
-        println!("\n");
-        ir_generation::run(
-            filter_config.include.as_ref(),
-            cli.verbose,
-            sway_ir::ExperimentalFlags {
-                new_encoding: run_config.experimental.new_encoding,
-            },
-        )
-        .instrument(tracing::trace_span!("IR"))
-        .await?;
-    }
-
-    Ok(())
-}
